Rewrite the gem into a database extension and change the name
I initially used the word "adapter" because the intention was to
create an actual adapter. Once I decided to just extend Sequel's
built-in adapters, I've still kept the name because the API still felt
like loading the adapter.

However, this name turned out cause confusion, especially for @headius
who was hoping this gem was an adapter for ActiveRecord that uses
Sequel, which would enable ActiveRecord users to leverage Sequel's JDBC
support and the JRuby team wouldn't need to maintain a separate adapter
for JDBC support.

To make the gem name accurate, we rename it to
"sequel-activerecord_connection", which makes it clear that it's an
extension for Sequel that integrates with an ActiveRecord connection.

We also turn it into a Sequel database extension, which is the standard
way of extending Sequel. Now people need to explicitly initialize the
correct adapter and load the extension. I think this explicitness is
beneficial because it removes the unnecessary magic, especially for
people who are already familiar with Sequel's API.

We also test that more types of Sequel exceptions are raise correctly.
This revealed that SQLite adapter wasn't raising correct exceptions for
exceptions like not-null or foreign-key constraint violations, because
(a) we didn't tell SQLite's database object to map additional exception
types, and (b) the ActiveRecord API wasn't calling SQLite methods Sequel
was expecting, which meant we had to call them manually.

We also set ActiveRecord's default timezone at the time of loading the
extension, so that we don't have to copy the current implementation of
the Database#timezone method. This means that folks who are overriding
ActiveRecord's default timezone after initializing Sequel will have to
do this for the Sequel object manually. Also, it seems ActiveRecord
doesn't support per-model timezone settings (as
`ActiveRecord::Base.default_timezone` is read in connection adapters),
so we should be fine on that front.