Improve reliability of transactions
We improve Database#_trans to always keep the transaction state in sync
when transactions are either added or removed when using
ActiveRecord::Base.transaction directly. This fixes several edge cases,
such transaction state not clearing after ActiveRecord connection is
finished.

This allows us to remove the hack we did in
Database#transaction_finished? that tried to remove Sequel transaction
state if the outer transaction was ActiveRecord's. This hack caused an
infinite loop in some cases, such as transactional Rails tests.

As part of this, we now make ActiveRecord automatically create a
savepoint when the enclosing Sequel transaction has `:auto_savepoint`
set to true. And vice versa, we make Sequel automatically create a
svaepoint when the enclosing ActiveRecord transaction has `joinable:
false` set.

We now also explicitly fail when the user tries to register a
Database#after_commit or Database#after_rollback for an ActiveRecord
transaction, instead of silently not executing them. We might support
this use case in the future with the `after_commit_everywhere` gem.