Require Sequel version 5.16.0 or greater
This commit changes savepoints from being booleans to being hashes,
which this gem is relying on:

https://github.com/jeremyevans/sequel/commit/673c80d7af0d2ebdcbc9938e0ade10cedc94ba84

This change was made in 5.16.0, so this gem won't actually work on Sequel
versions older than that. I think this is an acceptable restriction to
make, because I expect the vast majority of people using this gem will
be adding Sequel to their projects just now.