Improve general strategy for adapter integration
Previously, instead of letting Sequel execute the query, we were hooking
into the query execution chain and calling into Active Record to execute
the query instead. This meant we had automatic AR logging and other things
happening around query execution.

However, this approach actually had many flaws:

* Database#{copy_table,copy_into,listen} weren't working
* it didn't support prepared statements
* in SQLite integration we couldn't use it due to too many differences
* converting Active Record exceptions into Sequel exceptions was tricky
* we had to repeat Sequel adapter logic in some cases (e.g. JDBC)
* ...

The main issue was that we weren't letting Sequel's adapters to do their
thing. So we change the approach to let Sequel's adapter fully take
over, and add Active Record stuff around manually. The latter turned out
to be easy, because (a) we don't actually need locking, (b) lazy
transactions are disabled as soon as we fetch #raw_connection, and (c)
primary/replica support only needs to be handled for AR queries.

To fully support the Postgres adapter, we had to copy-paste the methods
from Sequel::Postgres::Adapter class (which is a subclass of
PG::Connection) and add them to AR's PG::Connection object. This ensures
Sequel::DatabaseDisconnectError is always properly raised, which isn't
so important as it's only used by Sequel's connection pool (which we
aren't using), but it might help us in handling prepared statements
better. Other adapter integrations have been simplified as part of this
change.

We also improved logging as part of this rewrite:

* label queries made by Sequel with "Sequel" in AR logs
* log all queries, including COPY and LISTEN
* log to Sequel logger(s) as well