Avoid using deprecated sqlite3 API
We were using some deprecated sqlite3 APIs, which was revealed after
switching on the verbose mode. The reason for that is that the Sequel
adapter relied on the default value of SQLite3::Database#results_as_hash
(which is nil), while Active Record's adapter sets it to true. We fix
that by temporarily setting the value to nil, which is safe because only
one thread should be able to access the raw connection object at a time.