Support prepared statements and bound variables
Each of the three Sequel adapters define #prepared_statements method on
the connection object in Database#connect, which holds the prepared
statements cache. Since in our case Database#connect is not being
called, we have to manually define these methods. This is all that's
needed to make prepared statements and bound variables work.