Log bound variables to Active Record logger
When using prepared statements or bound variables, Sequel will append
the variables array to the logged query.

Active Record's log subscriber has its own logic for logging bound
variables, but I didn't see an easy way to leverage it, as it seems to
rely a lot on Active Record internals, and it seems to require more
information than what we have (i.e. keys & values, whereas we have only
values).

To work around that, we simply append the variables to the SQL, just
like Sequel does for its own logs.