Clear statement cache on ActiveRecord::PreparedStatementCacheExpired
Active Record's transaction manager clears the statement cache on this
exception. We've tried to replicate it here, but it turns out the
`#after_failure_actions` method wasn't actually being called, because
it's a private method and we forgot to pass `true` as the second
argument to `#respond_to?`. Even if we change that, the statement cache
still wouldn't be cleared, because once we roll back the transaction,
`#current_transaction` will return a NullTransaction instead of
RealTransaction which `#after_failure_actions` checks.

We fix this by avoiding this method altogether and handling
ActiveRecord::PreparedStatementCacheExpired directly in the Postgres
integration, repeating what Active Record's transaction manager
does.