Use AR connection lock in Database#synchronize
Active Record's connection pool doesn't seem to guarantee that a single
connection can only be accessed by one thread at a time. In the commit
aadfb6e34d0a14bff2fa6e61b905e1e64621460d we've removed usage of AR
connection lock, which caused obscure exceptions on PG::Connection
methods on some cases (including segmentation faults), which seems to be
the result of the same connection being used by multiple threads at the
same time.

Even though the sole job of a connection pool is to disallow concurrent
access of the same connection, Active Record is still internally using
locks on the connection level, which indicates that concurrent access is
indeed possible. This is at least the case in Rails system tests, where
a single connection is shared across multiple threads, but it seems it
could be the case in development too.

So we add the same locking Active Record uses internally, and we do it
in Database#synchronize, which is the place where Sequel's connection
pool normally locks the connection to the current thread. The interlock
is necessary to allow for any autoloading during query execution.

Fixes #12