Support AR transaction callbacks with after_commit_everywhere
Previously, we couldn't support adding transaction hooks when Active
Record holds the transaction, or adding a savepoint hook when Active
Record holds any enclosing savepoint:

  ActiveRecord::Base.transaction do
    DB.after_commit { ... } # unsupported
  end

  DB.transaction do
    ActiveRecord::Base.transaction(requires_new: true) do
      DB.after_commit(savepoint: true) { ... } # unsupported
    end
  end

This is because we couldn't know whether an transaction was committed or
a savepoint was released if it was managed directly by Active Record.
Not supporting this pretty limiting, because e.g. it meant that Sequel
transaction hooks couldn't be used with Rails' transactional tests.

We address this by registering Active Record transaction callbacks
instead of Sequel hooks when it's required. Since generic after_commit /
after_rollback callbacks are not trivial to achieve with Active Record,
we use the handy after_commit_everywhere gem.

Active Record transaction callbacks have slightly different behaviour
than Sequel callbacks in some circumstances, but that should be
perfectly acceptable because users of this library should already be
familiar with Active Record's behaviour. In some cases this difference
is desirable, such as in transactional tests where after_commit callbacks
will get executed as soon as an enclosing savepoint is released, which
makes testing easier.

  ActiveRecord::Base.transaction(joinable: true) do
    DB.transaction do
      DB.after_commit { puts "after commit" } } } }
    end
  end
  # BEGIN
  # SAVEPOINT active_record_1
  # RELEASE SAVEPOINT active_record_1
  # after commit
  # COMMIT