Fix AR queries failing inside Sequel transactions with mysql2 adapter
Opening a Sequel transaction was overriding connection query options to
what Sequel expects, so when executing an Active Record query inside a
Sequel transaction, in most cases it would blow up because Active Record
expects different query options.

We fix this by moving overriding query options to Database#_execute,
which is the only place when Mysql2::Client#query method is called.
We'll likely do a similar thing for other adapters.