Allow using sql_log_normalizer DB extension
The sql_log_normalizer extension by default skips normalizing the SQL
passed to `#log_connection_yield` if there are no loggers set. Which
improves performance when SQL is not being logged.

In our case we don't have any loggers set, but we're delegating the
logging to Active Record, but we still want the received SQL string to
be normalized. So, we tell Sequel not to skip normalizing in this case.