Require Sequel 3.38+
Prior to Sequel 3.38, this extension would not automatically skip
testing the connection. This is because previously Sequel checked
original options hash for whether :test is set to true, not the
duplicated options hash set on the Sequel::Database object that is
available to extensions.

This behaviour could cause issues in Rails apps when the database
doesn't exist. For example, the `rails db:create` task initializes the
Rails application, but the database is expected not to exist beforehand.

In Sequel 3.38, connection testing has been moved into
Database#initialize, which checks :test option from the same hash object
as database options.