Update README.md
Why?
Specifying the exact location reduces the cognitive workload for those not new to the library, gets them flying almost immediately a "hello-world" example, almost without thinking.

I hope this is helpful.

I have added a note: re framework agnosticism.

If this is not useful, please feel free to close.