Patch PG type maps around prepared statements as well
Active Record uses custom type maps for results, while Sequel uses the
default type mapping, and performs its own type conversions. We were
bridging this discrepancy for normal queries, but this code wasn't being
called for prepared statements.

We fix this by moving the type map patch to a place called for both
normal and prepared statements. While here, we also fix a test that
ensures were not patching this around connection checkout, as it was
previously passing regardless.

Fixes #21