Only warn when a Sequel extension fails to initialize because there is no database
Rails applications are expected to be bootable without a database
created, so that rake tasks such as `db:create` can work properly.
Unfortunately, whereas Active Record is lazy with checking out
connections, some Sequel extensions eagerly check out a connection
to complete initialization.

Since this gem is used in applications where Active Record is the
primary ORM, I don't see value in exceptions being raised in Sequel when
there is no database, because that problem will bubble up eventually in
Active Record anyway. So, we instead only warn when a Sequel extension
failed to get loaded when the database is missing, allowing a Rails app
to still be bootable in this case.

Closes #27